made changes to older tests

+++ tests/BasicStarlinkTest.py
@@ -11,7 +11,7 @@ V = 4*np.pi*50*(R+600)**2 # volume of band
 S_i = [0]
 S_di = [0]
 D_i = [0]
-N_i = int(2.5e-8*V)
+N_i = 2.5e-8*V
 lam = 1000
 T = 50
 atmosphere = NCell([S_i], [S_di], [D_i], [N_i], [600], alt, [lam], min_lifetime=0)
